Added SwiftUIButtonProxy to the project.
SwiftUIButtonProxy class will expose our SwiftUI view (that we will
create in the next step) to React Native. In this commit we only expose
a Swift class to RN. In fact, this is the way how you expose Swift
components to React Native. Once you create a Swift file in your React
Native project, XCode will ask you if you'd like to create a bridging
header. Say "yes" as this header auto-bridges objc and swift realms.
Without it, we won't be able to return a SwiftUIProxy from the -(UIView
*)view {...} method.

P.S. Mind the "<ProjectName>-Swift.h" file in the Manager. It brings the
Swift realm to Objective-C.